refactor : #72 entity, model 패키지 model로 패키지명 통일
