+++ src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/comment/repository/v1/CommentRepository.kt
@@ -3,7 +3,7 @@ package sparta.nbcamp.gamenomeprojectserver.domain.comment.repository.v1
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
-import sparta.nbcamp.gamenomeprojectserver.domain.comment.entity.v1.Comment
+import sparta.nbcamp.gamenomeprojectserver.domain.comment.model.v1.Comment
 
 
 interface CommentRepository: JpaRepository<Comment, Long> {
