--- src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/reaction/entity/v1/ReactionType.kt
@@ -1,6 +1,0 @@
-package sparta.nbcamp.gamenomeprojectserver.domain.reaction.entity.v1
-
-enum class ReactionType {
-    Like,
-    DisLike
-}
