+++ src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/reaction/model/v1/ReactionType.kt
@@ -1,0 +1,6 @@
+package sparta.nbcamp.gamenomeprojectserver.domain.reaction.model.v1
+
+enum class ReactionType {
+    Like,
+    DisLike
+}
