+++ src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/reaction/repository/v1/ReactionRepository.kt
@@ -1,10 +1,10 @@
 package sparta.nbcamp.gamenomeprojectserver.domain.reaction.repository.v1
 
 import org.springframework.data.jpa.repository.JpaRepository
-import sparta.nbcamp.gamenomeprojectserver.domain.reaction.entity.v1.Reaction
-import sparta.nbcamp.gamenomeprojectserver.domain.reaction.entity.v1.ReactionType
+import sparta.nbcamp.gamenomeprojectserver.domain.reaction.model.v1.Reaction
+import sparta.nbcamp.gamenomeprojectserver.domain.reaction.model.v1.ReactionType
 
 interface ReactionRepository: JpaRepository<Reaction, Long> {
 
-    fun findByCommentIdAndReaction(commentId: Long, reaction: ReactionType):Reaction?
+    fun findByCommentIdAndReaction(commentId: Long, reaction: ReactionType): Reaction?
 }
