+++ src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/reaction/service/v1/ReactionService.kt
@@ -2,9 +2,9 @@ package sparta.nbcamp.gamenomeprojectserver.domain.reaction.service.v1
 
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import sparta.nbcamp.gamenomeprojectserver.domain.comment.entity.v1.Comment
-import sparta.nbcamp.gamenomeprojectserver.domain.reaction.entity.v1.Reaction
-import sparta.nbcamp.gamenomeprojectserver.domain.reaction.entity.v1.ReactionType
+import sparta.nbcamp.gamenomeprojectserver.domain.comment.model.v1.Comment
+import sparta.nbcamp.gamenomeprojectserver.domain.reaction.model.v1.Reaction
+import sparta.nbcamp.gamenomeprojectserver.domain.reaction.model.v1.ReactionType
 import sparta.nbcamp.gamenomeprojectserver.domain.reaction.repository.v1.ReactionRepository
 import sparta.nbcamp.gamenomeprojectserver.domain.user.repository.UserRepository
 import sparta.nbcamp.gamenomeprojectserver.exception.ModelNotFoundException
