--- src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/report/entity/v1/EntityType.kt
@@ -1,6 +1,0 @@
-package sparta.nbcamp.gamenomeprojectserver.domain.report.entity.v1
-
-enum class EntityType {
-    Review,
-    Comment
-}
