+++ src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/report/model/v1/EntityType.kt
@@ -1,0 +1,6 @@
+package sparta.nbcamp.gamenomeprojectserver.domain.report.model.v1
+
+enum class EntityType {
+    Review,
+    Comment
+}
