+++ src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/report/model/v1/Report.kt
@@ -1,4 +1,4 @@
-package sparta.nbcamp.gamenomeprojectserver.domain.report.entity.v1
+package sparta.nbcamp.gamenomeprojectserver.domain.report.model.v1
 
 import jakarta.persistence.*
 import org.hibernate.annotations.CreationTimestamp
