+++ src/main/kotlin/sparta/nbcamp/gamenomeprojectserver/domain/report/repository/v1/ReportRepository.kt
@@ -1,8 +1,8 @@
 package sparta.nbcamp.gamenomeprojectserver.domain.report.repository.v1
 
 import org.springframework.data.jpa.repository.JpaRepository
-import sparta.nbcamp.gamenomeprojectserver.domain.report.entity.v1.EntityType
-import sparta.nbcamp.gamenomeprojectserver.domain.report.entity.v1.Report
+import sparta.nbcamp.gamenomeprojectserver.domain.report.model.v1.EntityType
+import sparta.nbcamp.gamenomeprojectserver.domain.report.model.v1.Report
 
 interface ReportRepository : JpaRepository<Report, Long> {
     fun findByEntityType(entityType: EntityType): List<Report>
